# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ITIL v4 Change Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /role-assignments:
    get:
      operationId: listRoleAssignments
      tags: [roles]
      responses:
        "200":
          description: Assignments
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/RoleAssignment" }
    post:
      operationId: createRoleAssignment
      tags: [roles]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/CreateRoleAssignmentBody" } } }
      responses:
        "201":
          description: Created
          content: { application/json: { schema: { $ref: "#/components/schemas/RoleAssignment" } } }
  /role-assignments/{id}:
    delete:
      operationId: deleteRoleAssignment
      tags: [roles]
      parameters:
        - { in: path, name: id, required: true, schema: { type: integer } }
      responses:
        "204": { description: Deleted }

  # ---------- CHANGES ----------
...
  /changes/{id}/planning:
    get:
      operationId: getPlanning
      tags: [planning]
      parameters:
        - { in: path, name: id, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Planning
          content: { application/json: { schema: { $ref: "#/components/schemas/PlanningRecord" } } }
    put:
      operationId: upsertPlanning
      tags: [planning]
      parameters:
        - { in: path, name: id, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/UpsertPlanningBody" } } }
      responses:
        "200":
          description: Saved
          content: { application/json: { schema: { $ref: "#/components/schemas/PlanningRecord" } } }

  # ---------- TESTING / PIR ----------
  /changes/{id}/testing:
    get:
      operationId: getTesting
      tags: [testing]
      parameters:
        - { in: path, name: id, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Testing
          content: { application/json: { schema: { $ref: "#/components/schemas/TestRecord" } } }
    put:
      operationId: upsertTesting
      tags: [testing]
      parameters:
        - { in: path, name: id, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/UpsertTestingBody" } } }
      responses:
        "200":
          description: Saved
          content: { application/json: { schema: { $ref: "#/components/schemas/TestRecord" } } }
  /changes/{id}/pir:
    get:
      operationId: getPir
      tags: [testing]
      parameters:
        - { in: path, name: id, required: true, schema: { type: integer } }
      responses:
        "200":
          description: PIR
          content: { application/json: { schema: { $ref: "#/components/schemas/PirRecord" } } }
    put:
      operationId: upsertPir
      tags: [testing]
      parameters:
        - { in: path, name: id, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/UpsertPirBody" } } }
      responses:
        "200":
          description: Saved
          content: { application/json: { schema: { $ref: "#/components/schemas/PirRecord" } } }

  # ---------- APPROVALS ----------
...
components:
  schemas:
...
    RoleAssignment:
      type: object
      properties:
        id: { type: integer }
        roleKey: { type: string }
        userId: { type: integer }
        userName: { type: string }
        isDeputy: { type: boolean }
        primaryAssignmentId: { type: ["integer", "null"] }
      required: [id, roleKey, userId, userName, isDeputy]
    CreateRoleAssignmentBody:
      type: object
      properties:
        roleKey: { type: string }
        userId: { type: integer }
        isDeputy: { type: boolean }
        primaryAssignmentId: { type: ["integer", "null"] }
      required: [roleKey, userId, isDeputy]
...
    PlanningRecord:
      type: object
      properties:
        changeId: { type: integer }
        scope: { type: string }
        implementationPlan: { type: string }
        rollbackPlan: { type: string }
        riskAssessment: { type: string }
        impactedServices: { type: string }
        communicationsPlan: { type: string }
        successCriteria: { type: string }
        signedOff: { type: boolean }
        signedOffAt: { type: ["string", "null"], format: date-time }
        signedOffBy: { type: ["string", "null"] }
      required: [changeId, scope, implementationPlan, rollbackPlan, riskAssessment, impactedServices, communicationsPlan, successCriteria, signedOff]
    UpsertPlanningBody:
      type: object
      properties:
        scope: { type: string }
        implementationPlan: { type: string }
        rollbackPlan: { type: string }
        riskAssessment: { type: string }
        impactedServices: { type: string }
        communicationsPlan: { type: string }
        successCriteria: { type: string }
        signedOff: { type: boolean }
      required: [scope, implementationPlan, rollbackPlan, riskAssessment, impactedServices, communicationsPlan, successCriteria, signedOff]
    TestCase:
      type: object
      properties:
        name: { type: string }
        steps: { type: string }
        expectedResult: { type: string }
        actualResult: { type: string }
        status: { type: string, enum: [pending, passed, failed, blocked] }
      required: [name, steps, expectedResult, actualResult, status]
    TestRecord:
      type: object
      properties:
        changeId: { type: integer }
        testPlan: { type: string }
        environment: { type: string }
        testedBy: { type: ["string", "null"] }
        testedAt: { type: ["string", "null"], format: date-time }
        overallResult: { type: string, enum: [pending, passed, failed] }
        notes: { type: string }
        cases:
          type: array
          items: { $ref: "#/components/schemas/TestCase" }
      required: [changeId, testPlan, environment, overallResult, notes, cases]
    UpsertTestingBody:
      type: object
      properties:
        testPlan: { type: string }
        environment: { type: string }
        overallResult: { type: string, enum: [pending, passed, failed] }
        notes: { type: string }
        cases:
          type: array
          items: { $ref: "#/components/schemas/TestCase" }
      required: [testPlan, environment, overallResult, notes, cases]
    PirRecord:
      type: object
      properties:
        changeId: { type: integer }
        outcome: { type: string, enum: [successful, partial, failed, rolledback] }
        objectivesMet: { type: string }
        issuesEncountered: { type: string }
        lessonsLearned: { type: string }
        followupActions: { type: string }
        completedBy: { type: ["string", "null"] }
        completedAt: { type: ["string", "null"], format: date-time }
      required: [changeId, outcome, objectivesMet, issuesEncountered, lessonsLearned, followupActions]
    UpsertPirBody:
      type: object
      properties:
        outcome: { type: string, enum: [successful, partial, failed, rolledback] }
        objectivesMet: { type: string }
        issuesEncountered: { type: string }
        lessonsLearned: { type: string }
        followupActions: { type: string }
      required: [outcome, objectivesMet, issuesEncountered, lessonsLearned, followupActions]
...